requirements.txt
python-dotenv
google-generativeai
absl-py